Huge refactoring, split window class GUI into partial classes. Fixed tools not working properly on assembly reload.
